# [crypto/ed25519] ed25519 Concurrent Batch Verification + ZIP-215 Enforcement