--- a/codec/optional_packer.go
+++ b/codec/optional_packer.go
@@ -7,7 +7,7 @@ import (
 	"github.com/ava-labs/avalanchego/ids"
 	"github.com/ava-labs/avalanchego/utils/set"
 	"github.com/ava-labs/hypersdk/consts"
-	"github.com/ava-labs/hypersdk/crypto"
+	"github.com/ava-labs/hypersdk/crypto/ed25519"
 )
 
 // OptionalPacker defines a struct that includes a Packer [ip], a bitset
@@ -96,8 +96,8 @@ func (o *OptionalPacker) UnpackID(dest *ids.ID) {
 
 // PackPublicKey packs [pk] into OptionalPacker if [pk] is not an empty PublicKey.
 // Updates the bitset and offset accordingly.
-func (o *OptionalPacker) PackPublicKey(pk crypto.PublicKey) {
-	if pk == crypto.EmptyPublicKey {
+func (o *OptionalPacker) PackPublicKey(pk ed25519.PublicKey) {
+	if pk == ed25519.EmptyPublicKey {
 		o.skipBit()
 		return
 	}
@@ -107,11 +107,11 @@ func (o *OptionalPacker) PackPublicKey(pk crypto.PublicKey) {
 
 // UnpackPublicKey unpacks a PublicKey into [dest] if the bitset is set at
 // the current offset. Increments offset regardless.
-func (o *OptionalPacker) UnpackPublicKey(dest *crypto.PublicKey) {
+func (o *OptionalPacker) UnpackPublicKey(dest *ed25519.PublicKey) {
 	if o.checkBit() {
 		o.ip.UnpackPublicKey(true, dest)
 	} else {
-		*dest = crypto.EmptyPublicKey
+		*dest = ed25519.EmptyPublicKey
 	}
 }