--- a/crypto/ed25519/errors.go
+++ b/crypto/ed25519/errors.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2023, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package crypto
+package ed25519
 
 import "errors"